vpaasystem/requirements.txt
asgiref==3.11.0
Django==4.2.26
django-cors-headers==4.9.0
djangorestframework==3.16.1
mysqlclient==2.2.7
pillow==12.0.0
PyMySQL==1.1.2
sqlparse==0.5.4
tzdata==2025.2
